Remove an unused RemoteBuildServer interface
