+++ test/bll/notifications/notificationwatcherimpl.test.ts
@@ -8,22 +8,22 @@ import {Change} from "../../../src/bll/entities/change";
 import {Credentials} from "../../../src/bll/credentialsstore/credentials";
 import {InMemoryCredentialsStore} from "../../../src/bll/credentialsstore/inmemorycredentialsstore";
 import {NotificationWatcherImpl} from "../../../src/bll/notifications/notificationwatcherimpl";
-import {RemoteBuildServerImpl} from "../../../src/dal/remotebuildserverimpl";
 import {TeamCityOutput} from "../../../src/view/teamcityoutput";
 import {SummaryDao} from "../../../src/dal/summarydao";
 import {BuildDao} from "../../../src/dal/builddao";
+import {RemoteBuildServer} from "../../../src/dal/remotebuildserver";
 
 suite("Notification Watcher Implementation", () => {
 
     const EXPECTED_TIMEOUT = 400;
     test("When credentials absent it should sleep", function (done) {
-        const remoteBuildServerImpl: RemoteBuildServerImpl = getRemoteBuildServerWithSameEventCounter();
+        const remoteBuildServer: RemoteBuildServer = getRemoteBuildServerWithSameEventCounter();
         const summaryDaoImpl: SummaryDao = getPermanentSummaryDaoMock();
         const buildDaoImpl: BuildDao = getBuildDaoMock();
         const credentials: Credentials = undefined;
         const credentialsStore: InMemoryCredentialsStore = getCredentialsStoreMock(credentials);
         const teamCityOutput: TeamCityOutput = getOutputMock();
-        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServerImpl, summaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutput);
+        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServer, summaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutput);
 
         this.timeout(EXPECTED_TIMEOUT + 100);
         const timeout = setTimeout(done, EXPECTED_TIMEOUT);
@@ -34,7 +34,7 @@ suite("Notification Watcher Implementation", () => {
     });
 
     test("When credentials present it should try to get summary object", function (done) {
-        const remoteBuildServerImpl: RemoteBuildServerImpl = getRemoteBuildServerWithSameEventCounter();
+        const remoteBuildServer: RemoteBuildServer = getRemoteBuildServerWithSameEventCounter();
         const buildDaoImpl: BuildDao = getBuildDaoMock();
         const credentials: Credentials = new Credentials("URL", "user", "password", "userId", "sessionId");
         const credentialsStore: InMemoryCredentialsStore = getCredentialsStoreMock(credentials);
@@ -42,7 +42,7 @@ suite("Notification Watcher Implementation", () => {
 
         const mockedSummaryDaoImpl: SummaryDao = tsMockito.mock(SummaryDao);
         const summaryDaoImplSpy: SummaryDao = tsMockito.instance(mockedSummaryDaoImpl);
-        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServerImpl, summaryDaoImplSpy, buildDaoImpl, credentialsStore, teamCityOutput);
+        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServer, summaryDaoImplSpy, buildDaoImpl, credentialsStore, teamCityOutput);
 
         watcher.activate().then(() => {
             //do nothing
@@ -54,14 +54,14 @@ suite("Notification Watcher Implementation", () => {
     });
 
     test("If counter returns the same value it should sleep", function (done) {
-        const remoteBuildServerImpl: RemoteBuildServerImpl = getRemoteBuildServerWithSameEventCounter();
+        const remoteBuildServer: RemoteBuildServer = getRemoteBuildServerWithSameEventCounter();
         const summaryDaoImpl: SummaryDao = getPermanentSummaryDaoMock();
         const buildDaoImpl: BuildDao = getBuildDaoMock();
         const credentials: Credentials = new Credentials("URL", "user", "password", "userId", "sessionId");
         const credentialsStore: InMemoryCredentialsStore = getCredentialsStoreMock(credentials);
         const teamCityOutput: TeamCityOutput = getOutputMock();
 
-        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServerImpl, summaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutput);
+        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(remoteBuildServer, summaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutput);
 
         this.timeout(EXPECTED_TIMEOUT + 100);
         const timeout = setTimeout(done, EXPECTED_TIMEOUT);
@@ -72,7 +72,7 @@ suite("Notification Watcher Implementation", () => {
     });
 
     test("If there are new changes it should try to get extended info for build", function (done) {
-        const mutableRemoteBuildServerImpl: RemoteBuildServerImpl = getRemoteBuildServerWithNewEventCounter();
+        const mutableRemoteBuildServer: RemoteBuildServer = getRemoteBuildServerWithNewEventCounter();
         const mutableSummaryDaoImpl: SummaryDao = getMutableSummaryDaoMock();
         const credentials: Credentials = new Credentials("URL", "user", "password", "userId", "sessionId");
         const credentialsStore: InMemoryCredentialsStore = getCredentialsStoreMock(credentials);
@@ -82,7 +82,7 @@ suite("Notification Watcher Implementation", () => {
         tsMockito.when(mockedBuildDaoImpl.getById(tsMockito.anyNumber())).thenReturn(Promise.resolve(getSimpleBuild()));
         const buildDaoImplSpy: BuildDao = tsMockito.instance(mockedBuildDaoImpl);
 
-        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(mutableRemoteBuildServerImpl, mutableSummaryDaoImpl, buildDaoImplSpy, credentialsStore, teamCityOutput);
+        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(mutableRemoteBuildServer, mutableSummaryDaoImpl, buildDaoImplSpy, credentialsStore, teamCityOutput);
         watcher.activate().catch((err) => {
             done(new Error(`Unexpected error ${VsCodeUtils.formatErrorMessage(err)}`));
         });
@@ -93,7 +93,7 @@ suite("Notification Watcher Implementation", () => {
     });
 
     test("If there are new changes it should display new changes", function (done) {
-        const mutableRemoteBuildServerImpl: RemoteBuildServerImpl = getRemoteBuildServerWithNewEventCounter();
+        const mutableRemoteBuildServer: RemoteBuildServer = getRemoteBuildServerWithNewEventCounter();
         const mutableSummaryDaoImpl: SummaryDao = getMutableSummaryDaoMock();
         const buildDaoImpl: BuildDao = getBuildDaoMock();
         const credentials: Credentials = new Credentials("URL", "user", "password", "userId", "sessionId");
@@ -101,7 +101,7 @@ suite("Notification Watcher Implementation", () => {
         const mockedOutputImpl: TeamCityOutput = tsMockito.mock(TeamCityOutput);
         const teamCityOutputSpy: TeamCityOutput = tsMockito.instance(mockedOutputImpl);
 
-        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(mutableRemoteBuildServerImpl, mutableSummaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutputSpy);
+        const watcher: NotificationWatcherImpl = new NotificationWatcherImpl(mutableRemoteBuildServer, mutableSummaryDaoImpl, buildDaoImpl, credentialsStore, teamCityOutputSpy);
         watcher.activate().catch((err) => {
             done(new Error(`Unexpected error ${VsCodeUtils.formatErrorMessage(err)}`));
         });
@@ -145,14 +145,14 @@ suite("Notification Watcher Implementation", () => {
         return tsMockito.instance(mockedSummaryDaoImpl);
     }
 
-    function getRemoteBuildServerWithNewEventCounter(): RemoteBuildServerImpl {
-        const mockedRemoteBuildImplServer: RemoteBuildServerImpl = tsMockito.mock(RemoteBuildServerImpl);
+    function getRemoteBuildServerWithNewEventCounter(): RemoteBuildServer {
+        const mockedRemoteBuildImplServer: RemoteBuildServer = tsMockito.mock(RemoteBuildServer);
         tsMockito.when(mockedRemoteBuildImplServer.getTotalNumberOfEvents(tsMockito.anyString())).thenReturn(Promise.resolve(0), Promise.resolve(1), Promise.resolve(2));
         return tsMockito.instance(mockedRemoteBuildImplServer);
     }
 
-    function getRemoteBuildServerWithSameEventCounter(): RemoteBuildServerImpl {
-        const mockedRemoteBuildImplServer: RemoteBuildServerImpl = tsMockito.mock(RemoteBuildServerImpl);
+    function getRemoteBuildServerWithSameEventCounter(): RemoteBuildServer {
+        const mockedRemoteBuildImplServer: RemoteBuildServer = tsMockito.mock(RemoteBuildServer);
         tsMockito.when(mockedRemoteBuildImplServer.getTotalNumberOfEvents(tsMockito.anyString())).thenReturn(Promise.resolve(0));
         return tsMockito.instance(mockedRemoteBuildImplServer);
     }
